requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
braceexpand==0.1.7
charset-normalizer==2.1.1
click==8.1.3
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
import-expression==1.1.4
jishaku==2.5.1
line-profiler==4.0.2
multidict==6.0.3
python-dotenv==0.21.0
six==1.16.0
yarl==1.8.2
